Utkarsh/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic>=2.7.0
python-multipart==0.0.6
langchain==0.1.13
langchain-community==0.0.29
langchain-huggingface==0.0.1
huggingface-hub>=0.23.0
chromadb==0.5.0
sentence-transformers>=2.6.0
python-dotenv==1.0.0
requests==2.31.0
